src/main/java/com/github/brunomndantas/jscrapper/support/parser/array/text/primitive/ArrayPrimitiveByteTextParser.java
package com.github.brunomndantas.jscrapper.support.parser.array.text.primitive;

import com.github.brunomndantas.jscrapper.support.parser.array.text.ArrayTextParser;
import org.openqa.selenium.WebElement;

public class ArrayPrimitiveByteTextParser extends ArrayTextParser {

    public ArrayPrimitiveByteTextParser() {
        super(byte.class, (byte)0);
    }



    @Override
    protected Object parseElement(WebElement element) throws Exception {
        return Byte.parseByte(element.getText());
    }

}